+++ specs/artifacts_spec.py
@@ -1,31 +1,33 @@
+""" Tests for artifacts """
+
+import sys
 from flask import current_app
 from mamba import description, before, after, it
 from expects import expect, equal
-from doublex_expects import *
-from doublex import Mock
+#from doublex_expects import *
+#from doublex import Mock
 
-import sys
 sys.path.insert(0, 'specs')
 
-from specs.spec_helpers import Context
-from specs.mock_elasticsearch import ElasticMock
-
+#pylint: disable=wrong-import-position
+from specs.spec_helpers import Context, ElasticMock
+#pylint: enable=wrong-import-position
 
 with description('/artifacts') as self:
 
-    with before.each: 
+    with before.each:
         self.context = Context()
         current_app.es = ElasticMock()
 
     with after.each:
-            self.context.delete()
+        self.context.delete()
 
     with description('GET'):
         with before.each:
             current_app.es.mock(function_name="get",
-                            return_value={"_source":{"class_diagram.png":""}})
+                                return_value={"_source":{"class_diagram.png":""}})
             self.response = self.context.client().get("/artifacts/1")
-            
+
 
         with it('returns a 200 status code'):
             expect(self.response.status_code).to(equal(200))
@@ -33,14 +35,14 @@ with description('/artifacts') as self:
     with description('POST'):
         with before.each:
             current_app.es.mock(function_name="create",
-                            return_value={})
+                                return_value={})
 
             self.response = self.context.client().post("/artifacts",  json={
                 "id": "asdf",
                 "type": "image",
                 "tags": [],
                 "file_url": "asdf"})
-            
+
 
         with it('returns a 201 status code'):
             expect(self.response.status_code).to(equal(201))
